DjangoAdmin2/requirements.clip.txt
aiofiles==24.1.0
amqp==5.3.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.3.1
django-crontab==0.7.1
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.9.0.post1
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.0
fsspec==2024.12.0
ftfy==6.3.1
h11==0.14.0
holidays==0.52
huggingface-hub==0.27.0
idna==3.10
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
prophet==1.1.5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
stanio==0.5.0
sympy==1.13.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
yarg==0.1.9
